Исправление по результату проверки
