+++ MovieQuiz/Models/MostPopularMovie.swift
@@ -1,0 +1,31 @@
+//
+//  MostPopularMovie.swift
+//  MovieQuiz
+//
+//  Created by Igor on 05.09.2025.
+//
+
+import Foundation
+
+struct MostPopularMovie: Codable {
+    let title: String
+    let rating: String
+    let imageUrl: URL
+    
+    var resizedImageURL: URL {
+        let urlString = imageUrl.absoluteString
+        let imageUrlString = urlString.components(separatedBy: "._")[0] + "._V0_UX600_.jpg"
+        
+        guard let newURL = URL(string: imageUrlString) else {
+            return imageUrl
+        }
+        
+        return newURL
+    }
+    
+    private enum CodingKeys: String, CodingKey {
+        case title = "fullTitle"
+        case rating = "imDbRating"
+        case imageUrl = "image"
+    }
+}
