+++ MovieQuiz/Models/MostPopularMovies.swift
@@ -11,26 +11,3 @@ struct MostPopularMovies: Codable {
     let errorMessage: String
     let items: [MostPopularMovie]
 }
-
-struct MostPopularMovie: Codable {
-    let title: String
-    let rating: String
-    let imageUrl: URL
-    
-    var resizedImageURL: URL {
-        let urlString = imageUrl.absoluteString
-        let imageUrlString = urlString.components(separatedBy: "._")[0] + "._V0_UX600_.jpg"
-        
-        guard let newURL = URL(string: imageUrlString) else {
-            return imageUrl
-        }
-        
-        return newURL
-    }
-    
-    private enum CodingKeys: String, CodingKey {
-        case title = "fullTitle"
-        case rating = "imDbRating"
-        case imageUrl = "image"
-    }
-}
